m-ohm-core/requirements.txt
GPUtil==1.4.0
kafka==1.3.5
kafka_python==2.0.2
psutil==5.9.4
py_cpuinfo==9.0.0
